refactor: configs separated from services in other file
