+++ backend/internal/service/flag_sender/service.go
@@ -3,24 +3,28 @@ package flag_sender
 import (
 	"context"
 	"log/slog"
+	"time"
+
+	"github.com/jacute/prettylogger"
 )
 
 type storage interface {
+	GetConfigParameter(ctx context.Context, key string) (string, error)
 }
 
 type FlagSender struct {
-	log *slog.Logger
-	db  storage
-	cfg *config
-}
-
-type config struct {
+	log          *slog.Logger
+	db           storage
+	cfg          *config
+	shutdownChan chan struct{}
 }
 
 func New(log *slog.Logger, db storage) *FlagSender {
 	er := &FlagSender{
-		log: log,
-		db:  db,
+		log:          log,
+		db:           db,
+		cfg:          &config{},
+		shutdownChan: make(chan struct{}),
 	}
 	err := er.loadConfig(context.Background())
 	if err != nil {
@@ -30,12 +34,36 @@ func New(log *slog.Logger, db storage) *FlagSender {
 	return er
 }
 
-func (er *FlagSender) loadConfig(ctx context.Context) error {
-	const op = "service.flag_sender.LoadConfig"
-	log := er.log.With(slog.String("op", op))
+func (fs *FlagSender) Start() {
+	const op = "service.flag_sender.Start"
+	log := fs.log.With(slog.String("op", op))
+	log.Info("starting flag sender service")
 
-	_ = log
-	// TODO: imlement loadConfig
+	ticker := time.NewTicker(fs.cfg.submitPeriod)
+	for {
+		select {
+		case <-ticker.C:
+			// every tick load new config from db
+			err := fs.loadConfig(context.Background())
+			if err != nil {
+				log.Error(
+					"error reloading flag sender config from db",
+					prettylogger.Err(err),
+				)
+				continue
+			}
+			// TODO: implement flag sender logic
+		case <-fs.shutdownChan:
+			log.Info("shutting down flag sender service")
+			return
+		}
+	}
+}
 
-	return nil
+func (fs *FlagSender) Stop() {
+	const op = "service.flag_sender.Stop"
+	log := fs.log.With(slog.String("op", op))
+
+	log.Info("stopping flag sender service")
+	fs.shutdownChan <- struct{}{}
 }
